+++ lib/src/database/cursor/modern_cursor.dart
@@ -11,15 +11,21 @@ import 'package:mongo_dart/src/database/commands/aggreagation_commands/wrapper/c
 import 'package:mongo_dart/src/database/commands/aggreagation_commands/wrapper/change_stream/change_stream_operation.dart';
 import 'package:mongo_dart/src/database/commands/query_and_write_operation_commands/get_more_command/get_more_command.dart';
 import 'package:mongo_dart/src/database/commands/query_and_write_operation_commands/find_operation/find_operation.dart';
+import 'package:mongo_dart/src/database/commands/query_and_write_operation_commands/get_more_command/get_more_options.dart';
 import 'package:mongo_dart/src/database/utils/map_keys.dart';
 
 import '../../../mongo_dart.dart';
 
 typedef MonadicBlock = void Function(Map<String, dynamic> value);
 
+const defaultBatchSize = 101;
+
 class ModernCursor {
   ModernCursor(CommandOperation operation,
-      {bool? checksumPresent, bool? moreToCome, bool? exhaustAllowed})
+      {bool? checksumPresent,
+      bool? moreToCome,
+      bool? exhaustAllowed,
+      int? batchSize})
       : operation = operation,
         collection = operation.collection,
         db = operation.collection?.db ?? operation.db,
@@ -35,6 +41,15 @@ class ModernCursor {
     } else if (operation is ChangeStreamOperation) {
       isChangeStream = tailable = awaitData = true;
     }
+    var internalBatchSize = batchSize;
+    if (internalBatchSize == null) {
+      var operationBatchSize = operation.options[keyBatchSize] as int?;
+      if (operationBatchSize != null && operationBatchSize != 0) {
+        internalBatchSize = operationBatchSize;
+      }
+    }
+
+    _batchSize = internalBatchSize ?? defaultBatchSize;
   }
 
   /// This method allows the creation of the cursor from the Id and the
@@ -70,6 +85,7 @@ class ModernCursor {
     if (this.isChangeStream) {
       this.tailable = this.awaitData = true;
     }
+    _batchSize = defaultBatchSize;
   }
 
   State state = State.INIT;
@@ -80,6 +96,16 @@ class ModernCursor {
   bool tailable = false;
   bool awaitData = false;
   bool isChangeStream = false;
+
+  // Batch size for the getMore command if different from the default
+  late int _batchSize;
+  int get batchSize => _batchSize;
+  set batchSize(int _value) {
+    if (_value < 0) {
+      throw MongoDartError('Batch size must be a non negative value');
+    }
+    _batchSize = _value;
+  }
 
   // in case of collection agnostic commands (aggregate) is the name
   // of the collecton as returns from the first batch (taken from ns)
@@ -154,7 +180,7 @@ class ModernCursor {
   /// Convenience method for
   /// ```dart
   /// await nextObject();
-  ///  await close();
+  /// await close();
   /// ```
   Future<Map<String, Object?>?> onlyFirst() async {
     var ret = await nextObject();
@@ -167,8 +193,13 @@ class ModernCursor {
       return _getNextItem();
     }
 
+    var justPrepareCursor = false;
     Map<String, Object?>? result;
     if (state == State.INIT && operation != null) {
+      if (operation!.options[keyBatchSize] != null &&
+          operation!.options[keyBatchSize] == 0) {
+        justPrepareCursor = true;
+      }
       result = await operation!.execute();
       state = State.OPEN;
     } else if (state == State.OPEN) {
@@ -177,7 +208,9 @@ class ModernCursor {
         return null;
       }
       var command = GetMoreCommand(collection, cursorId,
-          db: db, collectionName: collectionName);
+          db: db,
+          collectionName: collectionName,
+          getMoreOptions: GetMoreOptions(batchSize: _batchSize));
       result = await command.execute();
     }
     if (result == null) {
@@ -196,6 +229,11 @@ class ModernCursor {
         cursorMap == null ? BsonLong(0) : BsonLong(cursorMap[keyId] ?? 0);
     // The result map returns last records while setting cursorId to zero.
     extractCursorData(result);
+    // batch size for "first batch" was 0, no data returned.
+    // Just prepared the cursor for further fetching
+    if (justPrepareCursor) {
+      return nextObject();
+    }
     if (items.isNotEmpty) {
       return _getNextItem();
     }